Split upscale server persistence
