+++ frontend/src/server/tools/upscale-constants.ts
@@ -1,0 +1,2 @@
+export const UPSCALE_TOOL_EVENT_NAME = 'tool_upscale_generate';
+export const UPSCALE_PLACEHOLDER_THUMB = '/assets/frames/thumb-1x1.svg';
