+++ frontend/src/server/tools/upscale.ts
@@ -1,33 +1,24 @@
 import { randomUUID } from 'crypto';
 import { ApiError, ValidationError } from '@fal-ai/client';
 import { getUpscaleToolEngine } from '@/config/tools-upscale-engines';
-import { isDatabaseConfigured, query, type QueryExecutor, withDbTransaction } from '@/lib/db';
+import { query } from '@/lib/db';
 import { getFalClient } from '@/lib/fal-client';
-import { getResultProviderMode } from '@/lib/result-provider';
-import { ensureBillingSchema } from '@/lib/schema';
 import { computeBillingProductSnapshot } from '@/lib/billing-products';
-import { getPlatformFeeCents } from '@/lib/pricing';
-import { getUserPreferredCurrency, type Currency } from '@/lib/currency';
-import { reserveWalletChargeInExecutor } from '@/lib/wallet';
-import { receiptsPriceOnlyEnabled } from '@/lib/env';
-import { isStorageConfigured, recordUserAsset, uploadFileBuffer, uploadImageToStorage } from '@/server/storage';
 import { buildStoredImageRenderEntries, resolveHeroThumbFromRenders } from '@/lib/image-renders';
+import { getPlatformFeeCents } from '@/lib/pricing';
+import { getUserPreferredCurrency } from '@/lib/currency';
+import { receiptsPriceOnlyEnabled } from '@/lib/env';
 import { ensureReusableAsset, upsertLegacyJobOutputs } from '@/server/media-library';
 import {
   UPSCALE_VIDEO_DYNAMIC_MARGIN_MULTIPLIER,
   clampUpscaleFactor,
   estimateImageUpscaleCostUsd,
   estimateVideoUpscaleCostUsd,
-  resolveUpscaleOutputFetchTimeoutMs,
   resolveUpscaleMode,
   resolveUpscaleOutputFormat,
   resolveUpscaleTargetResolution,
 } from '@/lib/tools-upscale';
 import type {
-  UpscaleMediaType,
-  UpscaleOutputFormat,
-  UpscaleToolEngineId,
-  UpscaleToolOutput,
   UpscaleToolRequest,
   UpscaleToolResponse,
 } from '@/types/tools-upscale';
@@ -40,15 +31,21 @@ import {
   cloneUpscalePricingWithDynamicTotal,
   extractUpscaleActualCostUsd,
   extractUpscaleOutput,
-  formatToImageMime,
-  formatToVideoMime,
   parseUpscaleRequestId,
   toUpscaleValidationMessage,
   type VideoMetadata,
 } from './upscale-request-utils';
+import { UPSCALE_PLACEHOLDER_THUMB, UPSCALE_TOOL_EVENT_NAME } from './upscale-constants';
+import { UpscaleToolError } from './upscale-errors';
+export { UpscaleToolError } from './upscale-errors';
+import {
+  createAtomicInitialUpscaleJob,
+  insertUpscaleToolEvent,
+  recordUpscaleRefundReceipt,
+  type PendingUpscaleReceipt,
+} from './upscale-job-persistence';
+import { persistUpscaleOutput } from './upscale-output-persistence';
 
-const TOOL_EVENT_NAME = 'tool_upscale_generate';
-const PLACEHOLDER_THUMB = '/assets/frames/thumb-1x1.svg';
 
 type RunUpscaleToolInput = UpscaleToolRequest & {
   userId: string;
@@ -69,371 +66,9 @@ export type RunUpscaleToolDependencies = {
   detectVideoMetadata?: (videoUrl: string, options?: { timeoutMs?: number }) => Promise<VideoMetadata | null>;
 };
 
-type PendingUpscaleReceipt = {
-  userId: string;
-  amountCents: number;
-  currency: string;
-  description: string;
-  jobId: string;
-  surface: typeof UPSCALE_SURFACE;
-  billingProductKey: string;
-  snapshot: PricingSnapshot;
-  applicationFeeCents: number | null;
-  vendorAccountId: string | null;
-};
-
-type CreateUpscaleInitialJobParams = {
-  userId: string;
-  jobId: string;
-  description: string;
-  amountCents: number;
-  currency: string;
-  billingProductKey: string;
-  pricingSnapshotJson: string;
-  applicationFeeCents: number | null;
-  vendorAccountId: string | null;
-  engineId: UpscaleToolEngineId;
-  engineLabel: string;
-  durationSec: number;
-  promptSummary: string;
-  settingsSnapshotJson: string;
-  preferredCurrency: Currency | null;
-};
-
-export class UpscaleToolError extends Error {
-  status: number;
-  code: string;
-  detail?: unknown;
-
-  constructor(message: string, options?: { status?: number; code?: string; detail?: unknown }) {
-    super(message);
-    this.name = 'UpscaleToolError';
-    this.status = options?.status ?? 500;
-    this.code = options?.code ?? 'upscale_tool_error';
-    this.detail = options?.detail;
-  }
-}
-
 function usdToCredits(value: number | null | undefined): number | null {
   if (typeof value !== 'number' || !Number.isFinite(value) || value <= 0) return null;
   return Math.max(1, Math.round(value * 100));
-}
-
-async function recordUpscaleRefundReceipt(receipt: PendingUpscaleReceipt, label: string, priceOnly: boolean) {
-  try {
-    await query(
-      `INSERT INTO app_receipts (
-         user_id,
-         type,
-         amount_cents,
-         currency,
-         description,
-         job_id,
-         surface,
-         billing_product_key,
-         pricing_snapshot,
-         application_fee_cents,
-         vendor_account_id,
-         platform_revenue_cents,
-         destination_acct
-       )
-       VALUES ($1,'refund',$2,$3,$4,$5,$6,$7,$8::jsonb,$9,$10,$11,$12)
-       ON CONFLICT DO NOTHING`,
-      [
-        receipt.userId,
-        receipt.amountCents,
-        receipt.currency,
-        label,
-        receipt.jobId,
-        receipt.surface,
-        receipt.billingProductKey,
-        JSON.stringify(receipt.snapshot),
-        priceOnly ? null : 0,
-        priceOnly ? null : receipt.vendorAccountId,
-        priceOnly ? null : 0,
-        priceOnly ? null : receipt.vendorAccountId,
-      ]
-    );
-  } catch (error) {
-    console.warn('[tools/upscale] failed to record refund receipt', error);
-  }
-}
-
-async function insertProvisionalUpscaleJob(
-  executor: QueryExecutor,
-  params: Omit<CreateUpscaleInitialJobParams, 'description' | 'preferredCurrency' | 'applicationFeeCents'>
-) {
-  await executor.query(
-    `INSERT INTO app_jobs (
-       job_id,
-       user_id,
-       surface,
-       billing_product_key,
-       engine_id,
-       engine_label,
-       duration_sec,
-       prompt,
-       thumb_url,
-       preview_frame,
-       status,
-       progress,
-       final_price_cents,
-       pricing_snapshot,
-       settings_snapshot,
-       currency,
-       vendor_account_id,
-       payment_status,
-       visibility,
-       indexable,
-       provisional
-     )
-     VALUES (
-       $1,$2,$3,$4,$5,$6,$7,$8,$9,$10,'pending',0,$11,$12::jsonb,$13::jsonb,$14,$15,'paid_wallet','private',FALSE,TRUE
-     )`,
-    [
-      params.jobId,
-      params.userId,
-      UPSCALE_SURFACE,
-      params.billingProductKey,
-      params.engineId,
-      params.engineLabel,
-      Math.max(1, Math.ceil(params.durationSec)),
-      params.promptSummary,
-      PLACEHOLDER_THUMB,
-      PLACEHOLDER_THUMB,
-      params.amountCents,
-      params.pricingSnapshotJson,
-      params.settingsSnapshotJson,
-      params.currency,
-      params.vendorAccountId,
-    ]
-  );
-}
-
-async function createUpscaleInitialJobInExecutor(
-  executor: QueryExecutor,
-  params: CreateUpscaleInitialJobParams
-): Promise<void> {
-  const reserveResult = await reserveWalletChargeInExecutor(
-    executor,
-    {
-      userId: params.userId,
-      amountCents: params.amountCents,
-      currency: params.currency,
-      description: params.description,
-      jobId: params.jobId,
-      surface: UPSCALE_SURFACE,
-      billingProductKey: params.billingProductKey,
-      pricingSnapshotJson: params.pricingSnapshotJson,
-      applicationFeeCents: params.applicationFeeCents,
-      vendorAccountId: params.vendorAccountId,
-      stripePaymentIntentId: null,
-      stripeChargeId: null,
-    },
-    { preferredCurrency: params.preferredCurrency }
-  );
-
-  if (!reserveResult.ok) {
-    if (reserveResult.errorCode === 'currency_mismatch') {
-      const lockedCurrency = (reserveResult.preferredCurrency ?? 'usd').toUpperCase();
-      throw new UpscaleToolError(`Wallet currency locked to ${lockedCurrency}. Contact support to request a change.`, {
-        status: 409,
-        code: 'wallet_currency_mismatch',
-        detail: { lockedCurrency },
-      });
-    }
-
-    throw new UpscaleToolError('Insufficient wallet balance for this upscale run.', {
-      status: 402,
-      code: 'insufficient_wallet_funds',
-      detail: {
-        balanceCents: reserveResult.balanceCents,
-        requiredCents: Math.max(0, params.amountCents - reserveResult.balanceCents),
-      },
-    });
-  }
-
-  await insertProvisionalUpscaleJob(executor, {
-    userId: params.userId,
-    jobId: params.jobId,
-    amountCents: params.amountCents,
-    currency: params.currency,
-    billingProductKey: params.billingProductKey,
-    pricingSnapshotJson: params.pricingSnapshotJson,
-    vendorAccountId: params.vendorAccountId,
-    engineId: params.engineId,
-    engineLabel: params.engineLabel,
-    durationSec: params.durationSec,
-    promptSummary: params.promptSummary,
-    settingsSnapshotJson: params.settingsSnapshotJson,
-  });
-}
-
-async function createAtomicInitialUpscaleJob(params: CreateUpscaleInitialJobParams): Promise<void> {
-  try {
-    await withDbTransaction(async (executor) => {
-      await executor.query(`SELECT pg_advisory_xact_lock(hashtext($1))`, [params.jobId]);
-      await createUpscaleInitialJobInExecutor(executor, params);
-    });
-  } catch (error) {
-    if (error instanceof UpscaleToolError) throw error;
-    throw new UpscaleToolError('Failed to save upscale job.', {
-      status: 500,
-      code: 'job_persist_failed',
-      detail: error instanceof Error ? error.message : error,
-    });
-  }
-}
-
-async function insertToolEvent(params: {
-  jobId: string;
-  engineId: UpscaleToolEngineId;
-  providerJobId?: string | null;
-  payload: Record<string, unknown>;
-}) {
-  if (!isDatabaseConfigured()) return;
-  try {
-    await ensureBillingSchema();
-    await query(
-      `INSERT INTO fal_queue_log (job_id, provider, provider_job_id, engine_id, status, payload)
-       VALUES ($1,$2,$3,$4,$5,$6::jsonb)`,
-      [
-        params.jobId,
-        getResultProviderMode(),
-        params.providerJobId ?? null,
-        params.engineId,
-        TOOL_EVENT_NAME,
-        JSON.stringify(params.payload),
-      ]
-    );
-  } catch (error) {
-    console.warn('[tools/upscale] failed to persist event log', error);
-  }
-}
-
-async function persistUpscaleOutput(params: {
-  output: UpscaleToolOutput;
-  mediaType: UpscaleMediaType;
-  userId: string;
-  jobId: string;
-  providerJobId?: string | null;
-  engineId: UpscaleToolEngineId;
-  engineLabel: string;
-  outputFormat: UpscaleOutputFormat;
-  durationSec?: number | null;
-}): Promise<UpscaleToolOutput> {
-  const { output, mediaType, userId, jobId, providerJobId, engineId, engineLabel, outputFormat, durationSec } = params;
-  if (!isStorageConfigured() || !output.url) return output;
-
-  try {
-    const controller = new AbortController();
-    const timeout = setTimeout(
-      () => controller.abort(),
-      resolveUpscaleOutputFetchTimeoutMs({ mediaType, durationSec })
-    );
-    let response: Response;
-    try {
-      response = await fetch(output.url, { signal: controller.signal });
-    } finally {
-      clearTimeout(timeout);
-    }
-    if (!response.ok) throw new Error(`fetch failed (${response.status})`);
-    const buffer = Buffer.from(await response.arrayBuffer());
-    if (!buffer.length) throw new Error('empty output');
-
-    if (mediaType === 'image') {
-      const mimeHeader = response.headers.get('content-type') ?? '';
-      const mime =
-        typeof output.mimeType === 'string' && output.mimeType.startsWith('image/')
-          ? output.mimeType
-          : mimeHeader.startsWith('image/')
-            ? mimeHeader
-            : formatToImageMime(outputFormat);
-      const upload = await uploadImageToStorage({
-        data: buffer,
-        mime,
-        userId,
-        prefix: 'upscale',
-        fileName: `upscale-${engineId}.${mime.split('/')[1] || 'png'}`,
-      });
-      const assetId = await recordUserAsset({
-        userId,
-        url: upload.url,
-        mime: upload.mime,
-        width: upload.width ?? output.width ?? null,
-        height: upload.height ?? output.height ?? null,
-        size: upload.size,
-        source: 'upscale',
-        metadata: {
-          originUrl: output.url,
-          jobId,
-          providerJobId: providerJobId ?? null,
-          tool: 'upscale',
-          engineId,
-          engineLabel,
-        },
-      });
-      return {
-        ...output,
-        url: upload.url,
-        width: upload.width ?? output.width ?? null,
-        height: upload.height ?? output.height ?? null,
-        mimeType: upload.mime,
-        originUrl: output.url,
-        assetId,
-        source: 'upscale',
-        persisted: true,
-      };
-    }
-
-    const mimeHeader = response.headers.get('content-type') ?? '';
-    const mime =
-      typeof output.mimeType === 'string' && (output.mimeType.startsWith('video/') || output.mimeType === 'image/gif')
-        ? output.mimeType
-        : mimeHeader.startsWith('video/') || mimeHeader === 'image/gif'
-          ? mimeHeader
-          : formatToVideoMime(outputFormat);
-    const upload = await uploadFileBuffer({
-      data: buffer,
-      mime,
-      userId,
-      prefix: 'upscale',
-      fileName: `upscale-${engineId}.${outputFormat === 'gif' ? 'gif' : outputFormat}`,
-    });
-    const assetId = await recordUserAsset({
-      userId,
-      url: upload.url,
-      mime,
-      width: output.width ?? null,
-      height: output.height ?? null,
-      size: buffer.length,
-      source: 'upscale',
-      metadata: {
-        originUrl: output.url,
-        jobId,
-        providerJobId: providerJobId ?? null,
-        tool: 'upscale',
-        engineId,
-        engineLabel,
-      },
-    });
-    return {
-      ...output,
-      url: upload.url,
-      mimeType: mime,
-      originUrl: output.url,
-      assetId,
-      source: 'upscale',
-      persisted: true,
-    };
-  } catch (error) {
-    console.warn('[tools/upscale] failed to persist output', {
-      engineId,
-      jobId,
-      error: error instanceof Error ? error.message : String(error),
-    });
-    return output;
-  }
 }
 
 export async function runUpscaleToolBase(
@@ -632,7 +267,7 @@ export async function runUpscaleToolBase(
       durationSec: videoMetadata?.durationSec ?? null,
     });
 
-    let thumbUrl = PLACEHOLDER_THUMB;
+    let thumbUrl = UPSCALE_PLACEHOLDER_THUMB;
     let renderIdsJson: string | null = null;
     let heroRenderId: string | null = null;
     if (mediaType === 'image') {
@@ -654,7 +289,7 @@ export async function runUpscaleToolBase(
       const renderEntries = buildStoredImageRenderEntries([outputWithThumb]);
       renderIdsJson = JSON.stringify(renderEntries);
       heroRenderId = outputWithThumb.url;
-      thumbUrl = resolveHeroThumbFromRenders([outputWithThumb]) ?? outputWithThumb.url ?? PLACEHOLDER_THUMB;
+      thumbUrl = resolveHeroThumbFromRenders([outputWithThumb]) ?? outputWithThumb.url ?? UPSCALE_PLACEHOLDER_THUMB;
       persistedOutput.thumbUrl = outputWithThumb.thumbUrl;
     }
 
@@ -720,12 +355,12 @@ export async function runUpscaleToolBase(
       console.warn('[tools/upscale] failed to persist media asset', { jobId }, assetError);
     });
 
-    await insertToolEvent({
+    await insertUpscaleToolEvent({
       jobId,
       engineId: engine.id,
       providerJobId: requestId,
       payload: {
-        event: TOOL_EVENT_NAME,
+        event: UPSCALE_TOOL_EVENT_NAME,
         userId: input.userId,
         status: 'success',
         engine: {
@@ -820,12 +455,12 @@ export async function runUpscaleToolBase(
       console.warn('[tools/upscale] failed to mark job as failed', updateError);
     }
 
-    await insertToolEvent({
+    await insertUpscaleToolEvent({
       jobId,
       engineId: engine.id,
       providerJobId,
       payload: {
-        event: TOOL_EVENT_NAME,
+        event: UPSCALE_TOOL_EVENT_NAME,
         userId: input.userId,
         status: 'error',
         engine: {
